refactor: s3 관련 타입 정의
