+++ src/pages/CreateInvitationPage.tsx
@@ -20,11 +20,7 @@ import {
 import useGalleryStore from '@/store/OptionalFeature/useGalleryFeatureStore';
 import { useOptionalFeatureStore } from '@/store/OptionalFeature/useOptionalFeatureStore';
 import useNoticeStore from '@/store/OptionalFeature/useNoticeFeatureStore';
-import {
-  InvitationDetail,
-  NoticeDetail,
-  S3UploadRequest,
-} from '@/types/invitationType';
+import { InvitationDetail, NoticeDetail } from '@/types/invitationType';
 import PreviewButton from '@/components/common/CreateInvitation/PreviewButton';
 import { useDebouncedInputStore } from '@/store/useDebouncedInputStore';
 import { useInvitationStore } from '@/store/useInvitaionStore';
@@ -32,6 +28,7 @@ import ReusableModal from '@/components/common/Modal/ReusableModal';
 import ChevronLeft from '@/components/icons/Chevron_LeftIcon';
 import useToast from '@/hooks/useToast';
 import Toast from '@/components/common/Toast';
+import { S3UploadRequest } from '@/types/s3Type';
 
 const sliceRanges = [[0, 3], [3, 13], [13]];
 // const AUTO_SAVE_MODAL_DURATION_MS = 3000;
