+++ src/types/s3Type.d.ts
@@ -1,0 +1,8 @@
+export interface S3UploadResponse {
+  imageUrls: string[];
+}
+
+export interface S3UploadRequest {
+  imageFiles: File[];
+  directory: string;
+}
